requirements.txt
streamlit
opensearch-py
python-dotenv
sentence-transformers
beautifulsoup4
requests
tiktoken
